examples/meeting-summarization/pyproject.toml
[project]
name = "meeting-summarization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-cpp-python>=0.3.16",
]
